pyproject.toml
[project]
name = "agentframe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "serpapi>=1.0.2",
    "tavily>=1.1.0",
]
